requirements.txt
choices==0.1
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
pytz==2018.9

